ISSUE-248 move TeamMailboxModule to new guice module
