+++ tmail-backend/apps/distributed/pom.xml
@@ -44,6 +44,10 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>team-mailboxes-guice</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>tmail-guice-jmap</artifactId>
             <type>test-jar</type>
             <scope>test</scope>
